Simplifying the logic for determining SerializationSettings
Removing duplicated and hard-to-read code that determined the C#
serialization/deserialization settings to be used for an IType. Changes
are needed because adding new checks for known primary types was failing
builds based cyclomatic complexity > 25.

This change adds extension methods that determine if an IType is either
a PrimaryType for a given KnownPrimaryType, or a
DictionaryType/SequenceType that contains values of that known type.