Adding logic for doing version comparison of Swagger specifications. (#1393)
* Transfered prototype version comparison code, with some rewrites, and added one single test.

* Added more tests

* Added tests.
Renamed some test data files.

* Fixed the processing of 'required' attribute.
Added some tests for 'required' changes.

* Added more tests related to parameter comparisons.

* Adding some more tests.
FIxing treatment of schemas that are, in fact, not referenced anywhere in the specification. Whether they have changed or not is uninteresting, since it has no impact on the wire format or on the client APIs.

* Adding tests for responses.
Adding tests for headers in parameters and responses.

* Adding tests of parameter and response constraints.

* Added tests related to more in-depth checking of 'enum' constraints.

* Invoke logic to validate the new Swagger document before bothering to compare it against the old version.
Adding commentary.